guidance-tag connections refactored
