+++ api-js/src/guidance-tag/objects/index.js
@@ -1,2 +1,3 @@
 export * from './guidance-tag'
+export * from './guidance-tag-connection'
 export * from './ref-links'
